fix: prevent newline amplification
A fuzzer found input sized 780kb, but the output becomes
180MB in size, ultimately taking 48s to parse on M1 Pro.

Note that the parsing itself is fast, but we seem to modify
the input in such a way that parsing takes way longer than it
should. That might be another issue which it will probably find
next, let's wait for it.