fix: Adjust `gix clean` warning and help for worktree fix
Since #1470 (for #1469), the repository's own nested worktrees are
no longer removed when using `gix clean`, even if they are nested
arbitarily under ignored directories. So no new mention of
worktrees (separate from "repositories") in the help and warning
message is accurate or required.

This removes that, which means really we are keeping the warning
message the same as it had been earlier, while still otherwise
bringing in new explanatory text in the `gix clean` options help.