Copy from target-specific build directories even on Windows
Changing the shell to bash *may* have been sufficient to cause
these directories to be used by successfully allowing environment
variables to be set and accessed using POSIX shell syntax. If so,
then this change together with that one *may* be enough either to
make it so that the 32-bit release archives really get 32-bit
executables on Windows or to reveal further reasons they do not.