Only use `cross` on Linux
This adds the `if:` key from the ripgrep workflow to the
"Use Cross" step, so that we only use `cross` instead of `cargo`
when both of the following hold:

(a) We are on a Linux runner.
(b) It is possible that we are cross-compiling.

Of those conditions, (a) is more important than (b), since `cross`
can be used even when not cross-compiling. However, on Windows, it
fails to create the environment, because there is no Docker image
available for it, and also it is a Windows environment so Windows
containers would have to be used, which are not available. This
would also not currently be able to work on Windows hosted GHA
runners, because they don't support nested virtualization.

Although using `cross` on Windows in this workflow had never worked
properly, there were two different incorrect behaviors:

- It was previously not actually being used at all, because
  PowerShell rather than bash was being used as the shell, and the
  attempt to use syntax like `>> $GITHUB_ENV` in PowerShell did not
  succeed at setting the `CARGO` environment variable (nor others).

  Although this is *conceptually* related to the bug where the
  32-bit (i686) Windows release archives actually contained 64-bit
  binaries (#1472), it does not seem to have been the direct cause,
  since cross-compiling with `cargo` should have worked. But the
  related failure to set the `$TARGET_FLAGS` and `$TARGET_DIR`
  environment variables was an important factor.
  This was one factor responsible for the bug where the 32-bit

- With bash used for all script steps on all operating systems, it
  failed early with an error from `docker` about not finding an
  image on ghcr.io corresponding to the target.

For some further details, see:
https://github.com/Byron/gitoxide/issues/1472#issuecomment-2254244666