Set $TARGET_FLAGS and $TARGET_DIR even without `cross`
This adds that step from the revised ripgrep workflow, moving the
commands into it. The new step is the same as in ripgrep except:

- Without `shell: bash`, because we set that for the whole job
  and do so on all operating systems.

- With a different quoting convention.

The related change of taking the executables from the specific
target directory has already been made. This is part of what is
needed to make that actually work.