Go back to always using `cross` on Linux
Because technically even the "non-cross" compilation is cross
compilation, as it is targeting `x86_64-unknown-linux-musl`.

When `cargo` rather than `cross` is used in that build, `ring`
fails to compile C code due to the `musl-gcc` command not being
present. Very likely other dependencies would also fail without
more tools installed on the system, and future features we wish to
support in binary releases would run into this in further ways.

An alternative approach here could be installing `musl-gcc` in a
prior step, much as we could also install an ARM gcc suite and
probably eliminate `cross` for linux-arm. For now, though, I'm
putting this back to using `cross` for all Linux builds.