Validate version against Cargo.toml
This is strongly inspired by an analogous check in the ripgrep
workflow, but done differently (from how that currently does it) by
parsing `Cargo.toml` using `yq`.

The `yq` command supports reading TOML (just not writing it). This
command is present on the `ubuntu-latest` runner.

This also moves the "Show the version" step into the new step, and
has that step also show the version extracted from `Cargo.toml`, so
as to make errors easier to understand and also to help verify that
this logic is really working.