Apparently we must specify the repo when not checked out
This adds a `REPOSITORY` environment variable in at the job level
for build-macos-universal2-release and references it in the two
steps that apparently need it.

Although these are the same two steps that use the token, that is
not added to steps that don't require it, since it is sensitive.