Check that macOS universal binaries are for the same features
As commented, this should catch if the features lists get out of
sync. (It may catch other problems too, but without this issue, the
check is likely not justified. Printing the list of artifacts at
the time publishing is about to occur may still be justified in
that case, though.)