Maybe slightly decrease risk of test precondition check failure
This canonicalizes the existing parent directory prefix of the
nonexistent-directory path used in `exe_info_tolerates_broken_tmp`,
so it may be less likely to be changed when obtained as the temp
dir path via `env::temp_dir()`.

That way, the test precondition check in the `set_temp_env_vars`
test helper is less likely to fail when setting up the test
succeeded. (The other caller already canonicalizes its path.)